added basic desgin
there's still alot to implement and to extend but this is basics
