+++ .flutter-plugins
@@ -1,8 +1,11 @@
 # This is a generated file; do not edit or check into version control.
-firebase_auth=C:\\Users\\Yaniv Galron\\AppData\\Roaming\\Pub\\Cache\\hosted\\pub.dartlang.org\\firebase_auth-0.15.5+3\\
-firebase_auth_web=C:\\Users\\Yaniv Galron\\AppData\\Roaming\\Pub\\Cache\\hosted\\pub.dartlang.org\\firebase_auth_web-0.1.2\\
-firebase_core=C:\\Users\\Yaniv Galron\\AppData\\Roaming\\Pub\\Cache\\hosted\\pub.dartlang.org\\firebase_core-0.4.4+3\\
-firebase_core_web=C:\\Users\\Yaniv Galron\\AppData\\Roaming\\Pub\\Cache\\hosted\\pub.dartlang.org\\firebase_core_web-0.1.1+2\\
-flutter_facebook_login=C:\\Users\\Yaniv Galron\\AppData\\Roaming\\Pub\\Cache\\hosted\\pub.dartlang.org\\flutter_facebook_login-3.0.0\\
-google_sign_in=C:\\Users\\Yaniv Galron\\AppData\\Roaming\\Pub\\Cache\\hosted\\pub.dartlang.org\\google_sign_in-4.4.1\\
-google_sign_in_web=C:\\Users\\Yaniv Galron\\AppData\\Roaming\\Pub\\Cache\\hosted\\pub.dartlang.org\\google_sign_in_web-0.8.4\\
+firebase_auth=D:\\platforms\\flutter\\.pub-cache\\hosted\\pub.dartlang.org\\firebase_auth-0.15.5+3\\
+firebase_auth_web=D:\\platforms\\flutter\\.pub-cache\\hosted\\pub.dartlang.org\\firebase_auth_web-0.1.2\\
+firebase_core=D:\\platforms\\flutter\\.pub-cache\\hosted\\pub.dartlang.org\\firebase_core-0.4.4+3\\
+firebase_core_web=D:\\platforms\\flutter\\.pub-cache\\hosted\\pub.dartlang.org\\firebase_core_web-0.1.1+2\\
+flutter_facebook_login=D:\\platforms\\flutter\\.pub-cache\\hosted\\pub.dartlang.org\\flutter_facebook_login-3.0.0\\
+google_sign_in=D:\\platforms\\flutter\\.pub-cache\\hosted\\pub.dartlang.org\\google_sign_in-4.4.1\\
+google_sign_in_web=D:\\platforms\\flutter\\.pub-cache\\hosted\\pub.dartlang.org\\google_sign_in_web-0.8.4\\
+shared_preferences=D:\\platforms\\flutter\\.pub-cache\\hosted\\pub.dartlang.org\\shared_preferences-0.5.6+3\\
+shared_preferences_macos=D:\\platforms\\flutter\\.pub-cache\\hosted\\pub.dartlang.org\\shared_preferences_macos-0.0.1+6\\
+shared_preferences_web=D:\\platforms\\flutter\\.pub-cache\\hosted\\pub.dartlang.org\\shared_preferences_web-0.1.2+4\\
