+++ lib/models/providerDetails.dart
@@ -1,0 +1,5 @@
+class ProviderDetails {
+  ProviderDetails(this.providerDetails);
+
+  final String providerDetails;
+}
