+++ lib/utils/storeTypes.dart
@@ -1,0 +1,1 @@
+enum StoreTypes { INT, STRING, BOOL }
