+++ pubspec.yaml
@@ -28,7 +28,7 @@ dependencies:
   flutter_signin_button: ^1.0.0
   flutter_facebook_login: ^3.0.0
   http: ^0.12.0+4
-
+  shared_preferences: ^0.5.6+3
 
 dev_dependencies:
   flutter_test:
